[Impeller] Make the default vertex buffer builder index size 16bit. (flutter/engine#36571)

+++ engine/src/flutter/impeller/renderer/vertex_buffer_builder.h
@@ -19,7 +19,7 @@
 
 namespace impeller {
 
-template <class VertexType_, class IndexType_ = uint32_t>
+template <class VertexType_, class IndexType_ = uint16_t>
 class VertexBufferBuilder {
  public:
   using VertexType = VertexType_;
